+++ lib/src/error/handler.dart
@@ -7,7 +7,20 @@ void errorHandler(Error t) {
 
   switch (t.type) {
     case ErrorType.SNF:
-      print('$prefixer Unable to find script named "${t.body['script']}".');
+      final String mainScript = t.body['script'];
+      final definitions =
+          (t.body['definitions'] as List).map((v) => v.toString()).toList();
+
+      print('$prefixer Unable to find script named "$mainScript".');
+      print('');
+
+      final bestMatch =
+          StringSimilarity.findBestMatch(mainScript, definitions).bestMatch;
+
+      if (bestMatch.rating >= 0.5) {
+        print('$prefixer Did you mean this?');
+        print('$prefixer    ${bestMatch.target}');
+      }
       break;
     case ErrorType.DNF:
       print('$prefixer Unable to find definitions.');